Add syntax and boilerplate code for pathrefs
This is to start with the path references implementation, which is a way
to make something like ${foo}/bin/ls fail at build time rather than at
runtime.

The operator I'm choosing here is => for now, but of course we can still
change that later.

Right now the implementation simply concatenates ${foo} with /bin/ls for
now until we have extended the Derive() arguments to cope with it.

We're going to check for pathrefs from the derivation referencing the
path rather than during evaluation time, because during the latter we
don't have a very good way to check for valid paths unless we use
read-write mode.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>