Turn string context into a structure
In order to pass in additional info about the path references, we need
to modify the type of string contexts a bit so it not only contains
plain store paths but also references to subpaths.

So for example ${foo => /some/path} and ${foo => /another/path} will
have a merged context with different pathrefs.

On the more high-level side we have its counterpart, which was PathSet
before. Unfortunately, we can't just modify the typedef, because it's
used not only for contexts but a variety of other things as well.

We now have a new typedef called Context, which is explicitly for these
contexts, as the name implies.

This commit shouldn't change functionality other than enlarging some
data structures to make room for the path refs.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>